chore(wmg): move child component to subfolder
