+++ frontend/src/components/BottomBanner/components/ModalContent/style.ts
