+++ frontend/src/components/BottomBanner/components/ModalContent/types.ts
